+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Support rust-analyzer as an alternate LSP server
 * Bump required VSCode version to 1.43, use language server protocol (LSP) v3.15
 
 ### 0.7.5 - 2020-05-06